Added some NotoUpdates in the source, also modified sphere_genericstone_functions.scp, sphere_guild_functions.scp and sphere_town_functions.scp to reflect noto changes when joining/leaving.

+++ 0.56.b/graysvr/CCharSpell.cpp
@@ -2985,7 +2985,7 @@ reflectit:
 int CChar::GetSpellEffect( SPELL_TYPE spell, int iSkillLevel, int iEffectMult )
 {
 	ADDTOCALLSTACK("CChar::GetSpellEffect");
-	long	iEffect = g_Cfg.GetSpellEffect( spell, iSkillLevel );
+	int	iEffect = g_Cfg.GetSpellEffect( spell, iSkillLevel );
 	return (iEffect * iEffectMult ) / 1000;
 }
